move perfzero benchmark to benchmark/
PiperOrigin-RevId: 305709689
