+++ official/benchmark/tfhub_memory_usage_benchmark.py
@@ -23,7 +23,7 @@ from absl import flags
 import tensorflow as tf
 import tensorflow_hub as hub
 
-from official.utils.testing.perfzero_benchmark import PerfZeroBenchmark
+from official.benchmark.perfzero_benchmark import PerfZeroBenchmark
 
 FLAGS = flags.FLAGS
